requirements.txt
azure-identity>=1.13.0
requests>=2.28.0
python-dotenv>=0.21.0
